+++ packages/falcon-front-kit/src/Locale/LocaleSwitcher.tsx
@@ -1,17 +1,8 @@
 import React from 'react';
 import { MutationFetchResult } from '@apollo/react-common';
 import { I18n } from '@deity/falcon-i18n';
-import { SetLocaleMutation, SetLocaleResponse, BackendConfigQuery } from '@deity/falcon-data';
-
-export const addCimodeLocale = (locales: string[]) => {
-  if (process.env.NODE_ENV === 'development') {
-    if (!locales.find(x => x === 'cimode')) {
-      locales.unshift('cimode');
-    }
-  }
-
-  return locales;
-};
+import { SetLocaleMutation, SetLocaleResponse } from '@deity/falcon-data';
+import { Locale } from '../Locale';
 
 export type LocaleItem = {
   code: string;
@@ -23,36 +14,27 @@ export type LocaleSwitcherRenderProps = {
   value: LocaleItem;
   items: LocaleItem[];
 };
-
 export type LocaleSwitcherProps = {
   children: (props: LocaleSwitcherRenderProps) => JSX.Element;
 };
-
 export const LocaleSwitcher: React.SFC<LocaleSwitcherProps> = ({ children }) => (
   <I18n>
     {(t, i18n) => (
       <SetLocaleMutation>
         {setLocale => (
-          <BackendConfigQuery>
-            {({ data: { backendConfig } }) => {
-              const { activeLocale } = backendConfig;
+          <Locale>
+            {({ locale, locales }) => {
+              const items = locales.map(code => ({ code, name: t(`languages.${code}`) }));
+              const value = { code: locale, name: t(`languages.${locale}`) };
 
-              const items = addCimodeLocale(backendConfig.locales).map(x => ({ code: x, name: t(`languages.${x}`) }));
-              const value = {
-                code: activeLocale,
-                name: t(`languages.${activeLocale}`)
-              };
-
-              const onChange = (x: LocaleItem) =>
-                setLocale({ variables: { locale: x.code } }).then(
-                  ({ data }: MutationFetchResult<SetLocaleResponse>) => {
-                    i18n.changeLanguage(data.setLocale.activeLocale);
-                  }
-                );
+              const onChange = ({ code }: LocaleItem) =>
+                setLocale({ variables: { locale: code } }).then(({ data }: MutationFetchResult<SetLocaleResponse>) => {
+                  i18n.changeLanguage(data.setLocale.activeLocale);
+                });
 
               return children && children({ items, value, onChange });
             }}
-          </BackendConfigQuery>
+          </Locale>
         )}
       </SetLocaleMutation>
     )}